requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
python-dotenv==1.0.0
geopy==2.4.1
feedparser==6.0.10
beautifulsoup4==4.12.2
numpy==1.24.4
pandas==2.1.3
schedule==1.2.1
python-multipart==0.0.6
httpx==0.25.2
